ff_draft_assistant/requirements.txt
pdfplumber
requests
openai
python-dotenv
"pymongo[srv]"
flask
espn-api
